audit 2

+++ backend/app/models/token_usage.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from sqlalchemy import ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -16,4 +18,4 @@ class TokenUsage(Base, CreatedAtMixin):
     feature: Mapped[int] = mapped_column(Integer, default=int(AIFeature.UNKNOWN))
     input_tokens: Mapped[int] = mapped_column(Integer)
     output_tokens: Mapped[int] = mapped_column(Integer)
-    estimated_cost: Mapped[float | None] = mapped_column(Numeric(12, 10), nullable=True, default=None)
+    estimated_cost: Mapped[Decimal | None] = mapped_column(Numeric(12, 10), nullable=True, default=None)
